[maint] Rename trial_stim to stim
