+++ lib/src/stim/stim_controller.dart
